各类统计数据查询存储逻辑重构

+++ pilipili-server-interact-service/src/main/java/com/bubble/pilipili/interact/pojo/dto/QueryCommentInfoDTO.java
@@ -62,9 +62,9 @@ public class QueryCommentInfoDTO implements Serializable {
     /**
      * 点赞数
      */
-    private Integer favorCount = 0;
+    private Long favorCount = 0L;
     /**
      * 点踩数
      */
-    private Integer dewCount = 0;
+    private Long dewCount = 0L;
 }
